Move /teensies code to /teensies/mine
